fix hack_stmt for non-repeated binding in sequence
Summary:
When building a repeat in sequence context:
```
vec[a, #{b*}, c]
```
make sure to treat non-repeat bindings as single values:
```
vec[#a, #{b*}, c]
```
that is - the two above statements should generate similar code.  Previously the macro was treating `#a` as a repeatable and trying to iterate over it (which, as a single value, it can't).

Reviewed By: oulgen

Differential Revision: D45533615

fbshipit-source-id: 38ec44847caf1ae62b71737a288fa71381225e1f