depgraph: remove transitive closure of extends edges, behind flag
Summary:
These are useless based on my understanding of the fanout algorithm.

That algorithm can be summarized this way:
* either a change affect a member (minor change), and then we follow edges starting from that member + sometimes transitively follow extends edges from the class.
* or it's a 'major change' and we always follow extends edges transitively.

In all cases, we never follow regular edges starting at classes, so those edges I'm removing here are useless.

I'm gating this behind a flag so as to be able to test my hypothesis with mass replay.

I'm also removing some rust unit tests because I don't want to have to update them each time I add or remove a flag.

Reviewed By: ljw1004

Differential Revision: D45483326

fbshipit-source-id: 99dc18397022d6d0b60986024723628e879992db