Fix performance regression on Shapes::toDict
Summary:
Recent diff D45395390 improved the typing of `Shapes::toDict` to account for the type of unknown fields, which is typically `supportdyn<mixed>` under implicit pessimisation. For a shape type such as `shape('a' => int, 'b' => NonSDT, _ => supportdyn<mixed>)`, the type of the dictionary values is computed as the union of the field types.

This can be expensive to compute for large number of fields. But in the common case, the known field types are subtypes of the unknown field type, so there is no need to include them in the union.

Differential Revision: D45533134

fbshipit-source-id: f553da431099d905f8d8e9cc52cc4b045cfe3d4e