Prefer && in folly::enumerate documentation
Summary:
When enabling `-Wrange-loop-bind-reference` Clang warns against binding a temporary to a const reference, but an rvalue reference is fine. Whether this is a useful warning is arguable, but in any case it's best to recommend the most general syntax.

The documentation was already recommending `&&` for the proxy iterator, but when using structured bindings it suggests `const&`. Update that to `&&` as well.

While at it, tweak the comment to make the structured bindings syntax more first-class.

Reviewed By: Gownta

Differential Revision:
D45566636
Ninja: This diff changes a comment, and arvr builds fail as usual

fbshipit-source-id: eeda73b902fde9f79e1af61285a79826c6c499c6