Don't optimize indirect jumps/calls into direct ones when the target is a stub
Summary:
The way this code works, this optimization can only be applied once.  After
that, the code pattern matching can't detect that we have a smashable
instruction.  It turns out that the smashable instructions were first being
smashed to point to a stub to later get a translation/retranslation for the
desired target.  As a result, the code would later fail to be smashed to point
to the desired target, and it would keep calling the stub and leaving the new
code unreachable.

Reviewed By: arnabde03

Differential Revision: D45554453

fbshipit-source-id: 40bbb9c07ff153ceffcdabbbe2ca868b3a4bf66d