add symbol index addenda to decl parse to avoid re-parse
Summary:
Add an optimization to avoid the re-parse necessary to update the "symbol index", which is what powers autocomplete and workspace symbol (the cmd+T or "#" search in VSCode).

The optimization relies on the fact that the decl parse we perform to produce `FileInfo` actually produces sufficient information for the symbol index already. Previously, we would perform an additional parse of each file using the "Fast Facts Parser" to build up the symbol index.

This optimization takes away a huge chunk of time. Here's an example execution where we have 16000 changes from the nearest loaded saved state and are processing all of them at once:
```
[2023-04-24 13:51:24.868] [ide-incremental] Actually processing 16029 changes...
[2023-04-24 13:51:28.066] [ide-incremental] Batch index completed in 3.197988 seconds
[2023-04-24 13:51:28.282] [ide-incremental] Updated reverse naming table in 0.079804 seconds
[2023-04-24 13:51:28.315] [ide-incremental] Updated forward naming table in 0.033499 seconds
[2023-04-24 13:51:28.316] [ide-incremental] Removed files from search index in 0.000054 seconds
[2023-04-24 13:51:51.738] [ide-incremental] Update search index with new files in 23.422453 seconds
```
Notices how the batch index completed in a bit over 3 seconds, but it took 23 seconds to build the search index with the old redundant sequential Fast Facts Parsing. With this diff, we save those 23 seconds.

# MISC
For the unrelated hhvm-flib-test failures:

Reviewed By: ljw1004

Differential Revision: D44993118

fbshipit-source-id: 1b4c4a3314818b07a1ea63afc474f270d99a90d8