Try to reach a fixed point for class constants when analyzing 86cinit
Summary:
If a class has a large number of constants defined in terms of each
other, then the analyze constants phase can require an excessive
number of rounds. This is because each round only processes each
constant once. So, for example, the first round will resolve one
constant, the second round will resolve the next (defined in terms of
the first), etc. Since each round has relatively high overhead, it's
advantageous to cut this down.

This is similar to other dependency chain problems we've already
encountered, and we can fix it the same way. When analyzing a 86cinit
function, we iterate until we reach a fixed point of the constant
types. This requires keeping local state about the already resolved
constants along with a worklist and a dependency map. This needs to be
threaded through the various places which might access a class
constant.

Reviewed By: kavoor

Differential Revision: D45545317

fbshipit-source-id: c2e95065a49e9ffbb08a2ba2c852e7b08bc988b4