Fix header dependencies in HHBBC
Summary:
type-system.h had a dependency on index.h, which meant that index.h
couldn't rely on Type being defined! As a result, we hacked around it
by using a template default argument trick. This is ugly. Make it so
that type-system.h doesn't require index.h. This requires moving a few
functions out of line. This means we can drop the template trick.

Reviewed By: oulgen

Differential Revision: D45574391

fbshipit-source-id: 1a37b96b2e9759000d3d6ca3fba6e7145d119262