exception message on TYPECHECKER_EXIT
Summary:
We recently had to revert a hack release because telemetry showed a huge spike in the number of TYPECHECKER_EXIT, specifically "Uncaught exception"

Unfortunately the event didn't say what the exception was! This diff addresses it. Now, the TYPECHECKER_EXIT exception will include the exception message and callstack.

Note: the mechanism by which this diff works is
1. the server has an uncaught exception within ServerTypeCheck
2. ServerTypeCheck.type_check uses `ServerUtils.with_exit_on_exception` to turn that into an `Exit.exit (Exit_status.Uncaught_exception)` https://www.internalfb.com/code/fbsource/[272d0144878c]/fbcode/hphp/hack/src/server/serverTypeCheck.ml?lines=1532
3. ServerMain has previously installed a hook by which all Exit.exit will write their information into a finale file. https://www.internalfb.com/code/fbsource/[272d0144878c]/fbcode/hphp/hack/src/server/serverMain.ml?lines=1441
4. MonitorMain, upon server death, attempts to read the finale file. It then writes what it knows into TYPECHECKER_EXIT event https://www.internalfb.com/code/fbsource/[272d0144878c]/fbcode/hphp/hack/src/monitor/monitorMain.ml?lines=677-679

I decided to write this diff via Exit.exit, and specifically by augmenting the finale file in the case of server exit, because it feels like a useful general-purpose mechanism that other things might benefit from too. E.g. if hh_client also wished to have a hook upon Exit.exit, then it its hook would get the richer information too.

Differential Revision: D45593274

fbshipit-source-id: 5887e54b35d7e12fb2e51d251f480e6398215843