Fix bugs that prevented vscode debugger from running async code
Summary:
While trying to understand why D45503309 works on hphpd but not on VSCode debug, I discovered that when one writes
```
await foo_async();
```
on the prompt,
vscode converts this to: `<?hh await foo_async();` which Hack parses as `Expr::Await`
while
hphpd converts this to: `<?hh (await foo_async());` which Hack parses as `Expr::Awaitall`

Async not allowed in regular functions check is only implemented for `Expr::Await` but not `Expr::Awaitall`.

So, first of all, lets implement this. Then we need to tell AST checkers that debugger code is special and it is inside a function to allow this await check to pass in debugger code as codegen will later put the expr into a function but this cannot be done earlier as it requires unnamed locals and declvars.

Reviewed By: aorenste

Differential Revision: D45612587

fbshipit-source-id: f3aa8b96455cc85a078ba96790303b6faba6d3b3