Fix JSON duplicate key leak test
Summary:
Calling JSON the first time allocates a bit of memory, so we
want to do that once first before setting the memory limit. This was
only sometimes making the test fail due to quirks about how the memory
manager track stats. Add a builtin to the intrinsics namespace to get
all of the memory manager stats. This has the side-effect of forcing
memory manager to sync it's record keeping. With this change the test
can now run remotely.

Reviewed By: aorenste

Differential Revision: D45673285

fbshipit-source-id: 58fdd5907a9714459db5ccbab50aadb608eb617b