Improve usability of the lexer API
Summary: The lexer API is a bit awkward to use because one needs to subclass `lex_handler`. However, we don't really need a full class because there is only one virtual method there that receives documentation comments. Replace it with `std::function` that defaults to ignoring comments. This allows passing lambdas which is much more convenient.

Reviewed By: iahs

Differential Revision: D45613326

fbshipit-source-id: 86aed4fc3c03e4e677610fad7176ded77259ebf2