(2nd attempt) [telemetry] report telemetry in CLIENT_CHECK_PARTIAL 2/2
Summary:
## About reverting

I'd first landed this diff D45469371. But the release went out, and lots of people found hh_server crashing with Uncaught_exception, and it got reverted D45546536. This is my second attempt.

The bug that caused a revert was simply that if you had hh_server running, and opened VSCode with HackIDE, then hh_server crashed! It was an oversight, not a deep or mysterious bug. Look at serverTypeCheck.ml:
```
let type_check (...) =
  if CheckKind.is_full_check kind then
    ServerProgress.ErrorsWrite.new_empty_file ...;
  let ... = type_check_unsafe in
  ServerProgress.ErrorsWrite.complete telemetry;

let type_check_unsafe (...) =
  let do_errors_file =
    genv.local_config.ServerLocalConfig.produce_streaming_errors
        && CheckKind.is_full
  in
  ...
  if do_errors_file then ServerProgress.ErrorsWrite.report new_errors;
  ...
  ServerProgress.ErrorsWrite.telemetry t;
  ...
```
The first version of this diff introduced `ServerProgress.ErrorsWrite.telemetry`, which writes a piece of telemetry to the streaming-errors file. However, it should have protected this write (as is done elsewhere in the code) with a check of `do_errors_file`. Without the check, it tried to write to the errors file at times when the errors file wasn't set up for writing (e.g. during a lazy check, or if the produce_streaming_errors flag were turned off), and this produced an exception.

This problem arose because IDE and hh_server are too tightly entangled, and we don't have many tests that exercise the two of them together. I think the right long-term answer is precisely what we're working on at the moment (and what this diff is in service of), namely to decouple IDE and hh_server entirely, so that dual tests aren't even needed. In the meantime, I reckon just continue with manual tests.

----------------------

## About this actual diff

The summary of D45468881 already set the stage: we want ServerTypeCheck to be able to incrementally write telemetry into the errors-file, which a client is able to accumulate, and log in the cast of Ctrl+C CLIENT_CHECK_PARTIAL.

That's what this diff achieves. It adds a new kind of thing that the errors-file can include, a "telemetry" object:
```
type errors_file_item =
  | Errors of {
      errors: Errors.finalized_error list Relative_path.Map.t;
      timestamp: float;
    }
  | Telemetry of {
      key: string;
      telemetry: Telemetry.t;
    }
```

It changes clientCheckStatus.ml so it can accumulate those telemetry items, and report them in CLIENT_CHECK_PARTIAL. (also in CLIENT_CHECK).

It changes serverTypeCheck.ml to report telemetry into the streaming-errors file about the size of the fanout, and typing_check_service.ml to report telemetry into the file about whether distc was used.

Reviewed By: nt591

Differential Revision: D45638328

fbshipit-source-id: 7ef143ad08c011ea42f387e1fb0fba359922747e