Add Eval.UnixServerRunPSPInBackground to run PSP actions in background
Summary:
This diff introduces `Eval.UnixServerRunPSPInBackground` which causes terminal control to be returned to the user prior to the start of PSP when cli-server is used to run scripts. The actions taken during PSP don't alter the return value of the command run and as a rule they should not perform any actions that the user would want to block on (they generally contain logging and cleanup operations).

When PSP runs in the background we disconnect stdout/stderr/stdin to avoid writing to the terminal after it has been returned, we also switch logging back to `error.log` so that any errors that occur during PSP can still be found in the error log.

This diff also fixes a bug where while running PSP in the foreground if the client is aborted the server can continue processing the request, this would occur because the monitor thread was shutdown prematurely before hack code had completed running. It also adds an `fsync` during server shutdown as I observed that output was not always fully flushed when the request completed.

Reviewed By: ricklavoie

Differential Revision: D45591071

fbshipit-source-id: 01618839d788402fa2c66024488423a4ebbf5162