update benchmark
Summary:
Updates existing benchmark and  add cpp benchmark to the TARGETS

Adds a benchmark susppender so the test measure the processing time of the passing messages between threads with in the requestpile excluding the initializing time

reorders the the producer and consumers so that consumers will be added before the producers. This is more realistic because a server will be configure to receive a load before anything starts to queue up.

Update the total number to 10,000 from 1000

Reviewed By: sazonovkirill

Differential Revision: D44905772

fbshipit-source-id: fb4e9d810ccf69a52f1e4afc812c211295851707