remove redundant validate_class_name in emit_class.rs
Summary:
Prior to this diffstack, hackc did some of its reserved-name-analysis in the parser, and some in emit_class.rs. The behavior of emit_class is that if there was a parser error, then it "fails fast" -- it doesn't go on to do its own reserved-name analysis as well.

The diffstack leading up to here has made it so the parser checks for every reserved name. Therefore, `emit_class`'s analysis for reserved names is redundant! This diff simply deletes it. Observe that no tests break, confirming that it's redundant.

Reviewed By: shayne-fletcher

Differential Revision: D45657513

fbshipit-source-id: 95e2afc876ef842f0e22e9d8222f16827bd79bd9