parser disallows reserved enum, enum-class names
Summary:
This is part of a suite of three diffs which move "reserved-name" checks into the parser. The diffs are all similar in implementation, but each one comes with different behavioral changes in hackc (hence why each has its own separate diff and separate justification).

1. D45638744 (prev diff) -- bring uniformity to classish (interface, class, trait)
2. **D45658285 (this diff) -- bring uniformity to enum and enum-class**
3. D45658761 (next diff) -- bring uniformity to alias (type, newtype, module newtype, case type, newctx)

This diff changes hackc behavior around enums. The question is which reserved type names may be used, and which ones can't?
```
enum <name> : int {}
enum class <name> : int {}
```

BEFORE: the following were errors in hh, but had various inconsistent behaviors in hackc
* names **bool, false, int, null, parent, string** -- these don't even parse as enums, for instance in `enum bool : int {}` the parser doesn't know what it's parsing and complains of missing semicolon
* names **callable, dynamic, float, mixed, nonnull, noreturn, nothing, num, resource, this, void, _** -- the parser accepts them, but emit_class reports an error.
* names **classname, darray, varray** -- these were allowed as enum names, even though not as class names
* names **N\arraykey, N\dynamic, N\float, N\mixed, N\nonnull, N\noreturn, N\nothing, N\num, N\resource, N\this, N\_** -- these were allowed as enum names within a namespace, but not at top level.

AFTER: every single one is now an error in hackc, just like hh, regardless whether it's in a namespace or not. The rules about reserved names for enum and enum-class are now identical to those for interfaces+traits+classes.

Is the behavior change okay, to disallow things that used to be accepted by hhvm? yes! Because the things were already rejected by hh.

----------

This diff also comes with two cosmetic changes.
1. Previously in hackc, some of the keywords had been rejected by the emitter and reported with single quotes while others had been rejected by the parser and reported with backticks. Now they are all rejected in the parser with backticks.
2. Previously in hh, some of the keywords were rejected by the parser `(Parsing[1002])` and others were rejected by hh `(Naming[2068])` and others were rejected by both and reported both errors. Now, there are no errors rejected solely by hh.

-------------

## How this diff works

1. In rust_parser_error.rs `enum_decl_errors` is used for checking enum and enum-class for parse errors. I added reserved-name-check to this function.
2. In emit_class.rs `validate_class_name` is used for checking classish (interface+trait+class) and also enum and enum-classes for reserved names. I left this code as is for now, but it's redundant since all errors that it might catch have already been reported by the parser. I will delete the redundant code a little further up the stack D45657513.

Reviewed By: oulgen

Differential Revision: D45658285

fbshipit-source-id: ab1d92f9d44d43bf23a6c5933a71cfb124585167