reword "as class name" -> "as a type name"
Summary:
This error text was bugging me:
```
interface arraykey {} // Cannot use `Arraykey` as class name as it is reserved
trait arraykey {} // Cannot use `Arraykey` as class name as it is reserved
class arraykey {} // Cannot use `Arraykey` as class name as it is reserved
```
We as hack developers know that they're all "classish", and the error report is saying the reserved name can't be used as a classish name. But this diff changes it to be more user friendly:
```
interface arraykey {} // Cannot use `Arraykey` as a type name as it is reserved
trait arraykey {} // Cannot use `Arraykey` as a type name as it is reserved
class arraykey {} // Cannot use `Arraykey` as a type name as it is reserved
```
This "as a type name" wording is the exact same wording as I introduced to report type alias reserved names. I did this because I think that it will be cleanest language design to simply have a single universal set of reserved names which apply to all type-names.

Reviewed By: edwinsmith

Differential Revision: D45675333

fbshipit-source-id: 1844c0b4064d11378447af84cd1803665a44f6ff