Optimize crc32
Summary:
X-link: https://github.com/facebook/folly/pull/1997

folly::crc32 implementation splits the buffer into unaligned head, multiple-of-16-bytes middle and unaligned tail. It uses boost to compute checksum for unaligned head and tail and SIMD for the aligned middle. We noticed that even when head and tail are empty (the buffer is fully aligned and its size is a multiple of 16), boost is still used and takes up > half of the profile.

{F984595200}

The optimization is to skip boost computation when tail is empty.

{F984596010}

Reviewed By: tonyxug, dmm-fb

Differential Revision: D45751314

fbshipit-source-id: 3b0fc5e72a0ca73cefc7ba427f2348c3aa14358c