Fix hover heisenbug (caching issue)
Summary:
In ServerHover we ask for the tast under dynamic assumptions, but other IDE stuff doesn't require it (or might not even want it).

BUT IDE stuff has caching. So when something else asks for a TAST first, ServerHover might get to see that.

Introduce some logic in the cache layer that makes the distinction between a TAST under dynamic assumptions and a TAST under "not" dynamic assumptions.

Only return the cached TAST if it was originally computed with the tco_under_dynamic_assumption true.

Reviewed By: andrewjkennedy

Differential Revision: D45777027

fbshipit-source-id: e4c35cf2da6fd3a0dcb48a743749ca526b99e0d0