Reinstate `validate_hint_habstr`
Summary:
The `validate_hint_habstr` pass was commented out saying that it didn't appear in the original naming passes - again, this is true but it _is_ a NAST check

There are also a couple of bugs in how to pass interacts with HKT elaboration pass due to fusion.

This pass is validating at the class / fundef / function / method level whilst the hkt elab was rewriting at the tparam level.

The elab pass will _delete_ HKTparams unless they are enabled.

Since tparams are always contained in one of class/fundef/func/meth, we can't get the ordering correct so we still see the params. The validation pass raises then errors on the parameters which we will later delete.

The fix is to rephrase the elab pass in terms of class/fundef/func/meth.

Note that we now raise unbound name errors when deleted HKTparams appear elsewhere in bounds. This is actually correct and an improvement on OCamk AFAICT

This is all a bit academic since we don't enable HKT anyway!

Reviewed By: shayne-fletcher

Differential Revision: D45725699

fbshipit-source-id: 0d732e5bfdc8ef63898965b18f0efe3322ee7241