Fix implementation of btrq Vasm instruction
Summary:
The s0 operand encodes the bit number to be tested, and it needs to be
converted into the corresponding bit mask that ARM's BIC instruction expects.
Also, BIC doesn't set the carry bit as specified by Vasm's btrq instruction,
so do a subtraction to set the carry bit accordingly.

Reviewed By: mdko

Differential Revision: D45673128

fbshipit-source-id: d3b59db567e55da0ee300eaf9768f2743128826c