fix broken reserved-name test
Summary:
There was a land race between
* 46aac1ecf11c - Wed May 10 11:13:39 2023 -0700 [[D45662224](https://www.internalfb.com/diff/D45662224)] Oguz Ulgen - Move setting the alias field of the type structure to hackc
* 38d65400ed20 - Wed May 10 15:49:01 2023 -0700 [[D45640535](https://www.internalfb.com/diff/D45640535)] Lucian Wischik - [hack] reserved keyword tests

The first diff altered the bytecode emitted for type aliases
```
type Foo = int;

.alias Foo = <"HH\int"> (2,2) """D:1:{s:4:\"kind\";i:1;}""";  // BEFORE
.alias Foo = <"HH\int"> (2,2) """D:2:{s:4:\"kind\";i:1;s:5:\"alias\";s:3:\"Foo\";}""";  // AFTER
```

The second diff had a test for the bytecode emitted for a range of keywords, expecting to see ".fatal" in many cases, but expecting to see success in a few. (and in those few success cases, the test unfortunately included the hackc output it expected to see).

DISCUSSION.
I don't know whether the hack/test/reserved tests are too exposed to bytecode-changes?
* I could leave them as they are, on the grounds that changes to bytecode for declarations are few and far between, and anyone who does change bytecode will understand how to run the "update" command that's shown in the terminal
* I could delete the "success" cases from the test suite
* I could alter the test so it merely greps for ".fatal" from the output of hackc, i.e. it only checks that the failures are exactly as we expect, and doesn't also check the successes.

Reviewed By: oulgen, edwinsmith

Differential Revision: D45778552

fbshipit-source-id: 2497941be3150d7de1308879bdf7af367388012c