Optimize builtins even when we are not inlining
Summary:
We don't need to only optimize builtins when we are inlining. IIUC we properly handle whether we get params by address or by value and the only place we were assuming we were inlining was in the `hphp_debug_caller_info` function, which I believe can be fixed by assuming we didn't find a non-skip-frame caller.

This is blocking some builtins in sandbox mode from being optimized.

Reviewed By: ricklavoie

Differential Revision: D45764952

fbshipit-source-id: 041ce8141fa16779fc842d8dc4dad08ae8255550