remove hackrs DeclParser -> GlobalOptions dependency
Summary:
I'd like to be able to use the hackrs family of crates for decl parsing
in the context of hackc decls and hhvm facts, without creating a dependency
on hh-centric GlobalOptions.

hackrs/decl_parser::DeclParser only needs one flag from GlobalOptions:
po_deregister_php_stdlib, and it's always false for the hackc/hhvm clients.

This refactors DeclParser to not depend on GlobalOptions.

Reviewed By: shayne-fletcher

Differential Revision: D45772881

fbshipit-source-id: 061e35733f50e1b3d9b5079f4d05e387906f1b17