io: do not recurse into directories on different FS
Summary:
In the case where a Watchman watch is created on a parent directory of an
EdenFS mount, Watchman will happily crawl the EdenFS repo as if it was on the
same filesystem as the root directory. However, this ends up creating a lot of
issues, starting with forcing EdenFS to fetch the entirety of the repository
and slowing it down as a result.

Since this is almost never the desired behavior, Watchman can simply stop
recursing on directories that are not on the same mount, solving the issue
above.

Reviewed By: chadaustin

Differential Revision: D45468181

fbshipit-source-id: 8117ba1e3980beeeadff32e50ff39f5c0b81a152