Change the way we traverse list when we drop connections
Summary:
Currently we are traversing list from the front to back, this means that we will first be dropping most active connections. In this diff changing this behaviour and traversing link backwards.

NOTE: Idea is that we want to drop connections with largest lifetime first.

## Details
First item we are starting with is `--idle_iterator`, **heuristically** this should give us connection with largest lifetime first.

NOTE: There is no guaranteed way that this approach will sort connections based on lifetime. It's totally possible we would have a connection which would periodically send some data and as a result will always end up on the beginning of the list.

Later if we decide to have life time based connection dropping (dropping connection whose lifetime exceeds certain threshold) it might make sense to traverse list from both sides but I believe currently this approach will be better than traversing list from the begging to end.

Reviewed By: jalopezsilva, dddmello

Differential Revision: D45649769

fbshipit-source-id: a7b5e4df48749361d5e62bffc0d32e5775b8c2e6