Recognise like-types in the TAST
Summary:
Union and intersection simplification can give us types that look like `(t1 & dynamic) | t2` which are more difficult to understand than `~t2 & t1`. Transform the former to the latter in the TAST, so that IDE hover types and other TAST consuming tools get the better types. (Implicit pessimisation introduces the types in some situations with partial enforcement e.g. enums as in the test case.)

Technically, we build `~t2 & (t1 | t2)` but in the cases where pessimisation introduces the ~ we have t2 <: t1 and so we end up with the simpler type

Reviewed By: madgen

Differential Revision: D45736546

fbshipit-source-id: 2c5212194924cd619ba06935674239c6de3f00e2