add SI to full index fallback
Summary:
To enable **workspace symbols** and **autocomplete** in clientIdeDaemon when  loading a non-saved state repository, we need additional information beyond the naming table. These should be the last IDE features that were previously unsupported when working with open source repositories.

Similar to the batch processing feature, we take advantage of the parsing work we already perform when indexing (building the naming table) to derive the extra "addenda" necessary to populate the "symbol index"; the symbol index is what we use to power the workspace symbols (searching for any top-level symbol in the repository, aka "cmd+T" on Mac VSCode) and autocomplete.

Note: I haven't rigorously tested concurrency behavior if two daemons on the same repository were to initialize in parallel. I don't think the `remove_file`-then-build should be a huge problem:  no matter how you re-organize it, one naming table build should succeed in the end? What could be a problem is if two builds occurred at the same time: if two builds tried to add the same symbols for the same files, would that count as a "collision"? We can avoid this situation entirely if we do the atomic rename, I think.

Reviewed By: nt591

Differential Revision: D45449679

fbshipit-source-id: 410449008d56d1b4d2ef35b29dd346901ff00456