Fix sync-signal race during forking
Summary:
If we're forked, and another process sends us a signal after we're set
up the signal handlers, but before the handling thread has started,
we'll just silently drop it. Fix this by starting the thread before
updating the signal handlers.

This was the cause of pcntl/tests/001.php flakiness, so we can bump
the timeouts back down.

Reviewed By: mdko

Differential Revision: D45873708

fbshipit-source-id: aa65f60a39e09a1fa374c24da99be7a5edbd9507