Improve definition docs
Summary:
Move interaction definition to a separate section. We already have one section per definition kind in all other cases even if there is some overlap as with structs and unions. This is more discoverable and user-friendly since one can navigate directly to services or interactions instead of figuring out that they are both defined in interfaces.

Rename `qualifier` to `field_qualifier` to avoid ambiguity with function qualifiers to be documented soon.

Improve wording of the Structs and Unions sections, making it more OSS-friendly, reducing the number of admonitions and adding links.

Change the field id upper bound from 65535 to the correct value of 32767.

Rename `FunctionSpecification` to `function` for consistency with other grammar symbols.

Reviewed By: iahs

Differential Revision: D45872361

fbshipit-source-id: 8be17f2899e1b185512ecf3a7fc275650bad29bf