Properly ignore bt_handler in stacktraces
Summary:
Having this function on the top of the stacktraces results in hphpcrash
aggregating a bunch of unrelated crashes in a single one, making it harder to
find the most common crashes and having crashbot trigger more often than it is
meant to.  It turns out we already tried to ignore bt_handler in the
stacktraces, but it was not working due to a typo.

Reviewed By: jtwarren, binliu19

Differential Revision: D45877062

fbshipit-source-id: 5c0c9e8601137c9e718180432e837b9d836a9d7a