Expose Addr() method in ServerTransport interface
Summary:
There is currently no way to get the address (i.e. port) on which a Go server is listening.

This is especially inconvenient for scenarios when a server is started on "127.0.0.1:0" - i.e. pick any port.

Expose the method via ServerTransport interface, so that callers can retrieve the address being listened on.

The actual implementations already exist, they are just not exposed via the interface:
https://www.internalfb.com/code/fbsource/[12903162ad5a]/fbcode/thrift/lib/go/thrift/server_socket.go?lines=95
https://www.internalfb.com/code/fbsource/[12903162ad5a]/fbcode/thrift/lib/go/thrift/ssl_server_socket.go?lines=89

I need this to unblock Go-Thrift conformance tests downstream: D44157004

Reviewed By: abulimov

Differential Revision:
D45885994

Privacy Context Container: L1158612

fbshipit-source-id: 583c36060551b8fd06b024fa49c17aebca701f07