Fix guard in cli-server
Summary: This change was supposed to be push-safe with old servers that did not yet support the "background" command, however, if the server issued an "exit" without first performing a "background" the `finish` guard in `run_command_on_cli_server` would never run as we would hit `exit` and not complete a teardown of the current frame. This would cause the foreground process on the client to see a hangup from the background process but not receive an exit code.

Reviewed By: ricklavoie

Differential Revision: D45914688

fbshipit-source-id: 7cf4111b714ee8e46933b1ebc6406da61d1dd8a2