add frame length parser
Summary:
Currently the logic for parsing a Rocket from the AsyncTransport has multiple different parsing strategies are in the same logic space (Parser.h and Parser-inl.h).  In these files there are around 4 different implemetations of paring logic controlled with if/else statements. There are also different different buffers that can be used depending on the logic being used. This diff creates a ParserStrategy class that can be implemented to encapsulate the different parsing strategies that we want to use. The idea is to eventually move the Parser logic in to different strategy classes, and remove the Parser-inl.h file.

There's an implementation of the ParserStrategy that uses the RSocket frame length to determine when to emit a new frame that shows you'd use the ParserStrategy.

Reviewed By: praihan, AkramaMirza

Differential Revision: D43428840

fbshipit-source-id: bd73f24c3adc4a3545ef7e655c0ebf0d3612d013