Housekeeping: rename struct Facts -> FactsExtension
Summary:
we use `namespace Facts` all over the place, so lets not also have
a `struct Facts` to confuse things. Most other extensions use the
naming convention `FooExtension`, so go with that.

Drop a few unnecessary `public` qualifiers.

AutoloadMap::Result was only used in AutoloadHandler code and only
has one bit of state anyway, so replace it with `bool`.

Reviewed By: paulbiss

Differential Revision: D45909272

fbshipit-source-id: 404f13db959b887b5c0d62ba34dc6ef864c73e38