Print out address, number of elements, refcount of Array types
Summary:
Quick fix while still we finishing implementing printing the elements of Arrays in LLDB. We should not print out the contents as "None" (as is done right now for unimplemented/unhandled) -- this is misleading. This fixes the following, e.g.:
```
(HPHP::Array) {
  m_arr = None
}
```

to instead print something like
```
(HPHP::Array) (0x6060008122a0) ArrayData[Vec]: 4 element(s) refcount=2
        { (Showing elements not yet implemented) }
```

Note: for every variable, you can use `dwim-print -R -- v` (or `expression -R -- v`) to print it out without the pretty printers, in case there are issues in the future.

In the next diff in the works, I'll add printing the elements, and update the newly added test cases.

Reviewed By: oulgen

Differential Revision: D45932994

fbshipit-source-id: 1885732bbd4ece3705750fbec9cf7cfa82fbe01b