New constructor to add quickfixes when evaluating `Reasons_callback.t`s
Summary: There are use cases where we want to add a `Pos.t Quickfix.t` when we have a `Reasons_callback.t` in hand. Currently, those use cases are stashing the quickfix in the `Secondary.t` error to which the callback is applied. This is not the intended use of secondary errors so we add a combinator which allows modification of the callback to contain the quickfixes

Reviewed By: mheiber

Differential Revision: D45947487

fbshipit-source-id: 061dfbfba42559b226f7e05123685079bd8493a1