Implement "Extract into Method" refactoring
Summary:
Steps to use:
- select some code that you want to extract into a method
- cmd-'.' or click the lightbulb
- select "Extract into method"

Example:
https://pxl.cl/2FHsS

## Advice for reviewing

As shown in the large test suite, there is a lot of behavior here, including:
- handling of plain functions, iterators, async functions, async iterators
- modifiers, parameters, and return type for the extracted function
- whether an expression or a set of statements is selected

To play with this feature, I recommend opening up some real code, and pasting in some of the examples from the test suite.

In the implementation, the code is split into two main parts:
- find the information needed to provide the refactoring
- generate the refactoring–I recommend reading this first, since otherwise it may not be obvious what all the information collected is *for*

## Next

the meta-only section below gives more context for these changes and follow-on work.

Reviewed By: ljw1004

Differential Revision: D44757198

fbshipit-source-id: 9a94d2a3b5ddf8765e26c9ca100877c8affaaf6a