refactor "flip around comma" to separate edit resolution.
Summary:
refactor the "flip around comma" code action to separate the logic for generating edits from the logic for finding candidate code actions.

This is a noop change that imo increases clarity. My goal is to makes it easier to show what lazy code action resolution is like later in the stack (D45828606)

Reviewed By: ljw1004

Differential Revision: D45852909

fbshipit-source-id: b61039357e6e6b4701b6b2a446db60fce95035a3